requirements.txt
tqdm==4.51.0
numpy==1.19.3
torch==1.7.0
torchvision==0.8.1
albumentations==0.5.2
Pillow==8.1.2
scikit_learn==0.24.1
#efficientnet_pytorch==0.7.0